Implement passenger count tracking and verification in backup_location.php
- Added passenger count update functionality with server-side validation
- Implemented client-side passenger count management with add/remove buttons
- Enhanced device verification process to secure passenger count updates
- Added new route for passenger count updates in backup_location.php
- Improved UI with responsive passenger count controls
- Integrated session-based device verification for passenger count modifications

+++ admin/fetch_data.php
@@ -56,7 +56,7 @@ if ($result->num_rows > 0) {
                     <a href='edit_jeepney.php?ids={$row['ID']}&id={$ids}' class='btn-icon edit' title='Edit'>
                         <i class='bx bxs-edit'></i>
                     </a>
-                    <a href='delete_jeepney.php?ids={$row['ID']}&id={$ids}' class='btn-icon delete' title='Delete'>
+                    <a href='delete_jeepney.php?ids={$row['ID']}&id={$ids}' class='btn-icon delete' title='Delete' onclick='return confirm(\"Are you sure you want to delete this account?\");'>
                         <i class='bx bxs-trash'></i>
                     </a>
                 </div>
